# [Kernel][Triton][AMD] Change default block size for triton_scaled_mm to 128 for 3-5x speedup