--- a/vllm/model_executor/layers/quantization/compressed_tensors/triton_scaled_mm.py
+++ b/vllm/model_executor/layers/quantization/compressed_tensors/triton_scaled_mm.py
@@ -126,9 +126,9 @@ def triton_scaled_mm(input: torch.Tensor,
                      scale_b: torch.Tensor,
                      out_dtype: Type[torch.dtype],
                      bias: Optional[torch.Tensor] = None,
-                     block_size_m: int = 128,
-                     block_size_n: int = 128,
-                     block_size_k: int = 128) -> torch.Tensor:
+                     block_size_m: int = 32,
+                     block_size_n: int = 32,
+                     block_size_k: int = 32) -> torch.Tensor:
     M, K = input.shape
     N = weight.shape[1]
 
@@ -151,6 +151,22 @@ def triton_scaled_mm(input: torch.Tensor,
     result = torch.empty((M, N), dtype=out_dtype, device=input.device)
 
     has_scalar = lambda x: x.shape[0] == 1 and x.shape[1] == 1
+
+    is_small_N = N < 8192
+    next_power_of_2_M = max(32, triton.next_power_of_2(M))
+    if next_power_of_2_M <= 32:
+        if is_small_N:
+            shape = (64, 64, 256)
+        else:
+            shape = (64, 128, 256)
+    elif next_power_of_2_M <= 64:
+        shape = (64, 64, 256)
+    elif next_power_of_2_M <= 128:
+        shape = (64, 128, 128)
+    else:
+        shape = (128, 128, 128)
+
+    block_size_m, block_size_n, block_size_k = shape
 
     block_size_sa = 1 if has_scalar(scale_a) else block_size_m
     block_size_sb = 1 if has_scalar(scale_b) else block_size_n