Changed default block-size for triton_scaled_mm to 128x128x128 from 32x32x32 for better performance.  This results in roughly 3-5x speedup.

python benchmarks/benchmark_latency.py --dtype bfloat16 --enable-chunked-prefill False --load-format dummy --batch-size 64 --num-iters-warmup 2 --num-iters 5 --input-len 2048 --output-len 128  --model /models/Phi-3-medium-128k-instruct-quantized.w8a8/

Before:

Avg latency: 14.48 seconds

After:

Avg latency: 5.52 seconds

python benchmarks/benchmark_throughput.py --dtype bfloat16 --enable-chunked-prefill False --load-format dummy --input-len 2048 --output-len 128  --model /models/Phi-3-medium-128k-instruct-quantized.w8a8/

Before:

Throughput: 10269.32 tok/s

After:

Throughput:  31150.8 tokens/s

FIX #xxxx (*link existing issues this PR will resolve*)

**BEFORE SUBMITTING, PLEASE READ https://docs.vllm.ai/en/latest/contributing/overview.html **