@rasmith 2025-01-07
> This is an impressive improvement! Could you also show comparisons for equal input len/output len workloads, preferably with low batchsize? This could regress the TPOT for small decode batches.
> 
> It seems there is no tuning for this kernel at the moment, so maybe this could benefit from a simple heuristic for the extreme problem sizes or a few `@triton.autotune` configs for the blocksizes.
 
@mgoin 
When just using 128x128x128 it gave better performance from some, but not all.  So, I used the heuristic from here: https://github.com/rasmith/vllm/blob/187e32997cdc20bbed5c21d3cef2609ab8ed9080/csrc/quantization/cutlass_w8a8/scaled_mm_c3x_sm90_int8_dispatch.cuh#L116.

I ran across various models and configs and was able to get improvement for all of the configs I tried.  Average speedup is ~ 2.8x.